Rename MailIndexer to WebMailIndexer
