+++ test/tests/Suite.ts
@@ -79,7 +79,7 @@ import "./api/worker/search/EventQueueTest.js"
 import "./api/worker/search/IndexUtilsTest.js"
 import "./api/worker/search/IndexerCoreTest.js"
 import "./api/worker/search/IndexedDbIndexerTest.js"
-import "./api/worker/search/MailIndexerTest.js"
+import "./api/worker/search/WebMailIndexerTest.js"
 import "./api/worker/search/IndexedDbMailIndexerBackendTest.js"
 import "./api/worker/search/IndexedDbSearchFacadeTest.js"
 import "./api/worker/search/SearchIndexEncodingTest.js"
